* Fix a few warnings ubuntu's g++ detected * Update bigsquares correctly.... (fixes crashes)
git-svn-id: https://bloopz.zapto.org/svn/nightfall/trunk@1094 d57568f3-fe0c-0410-b608-e86116ac1d00

+++ src/unitrender.cpp
@@ -578,7 +578,7 @@ namespace Game
 
 		void RenderProjectile(Projectile* proj)
 		{
-			if (SquareIsVisible(currentPlayerView, proj->pos.x, proj->pos.y))
+			if (SquareIsVisible(currentPlayerView, (int) proj->pos.x, (int) proj->pos.y))
 			{
 				Model* model = proj->type->model;
 				glPushMatrix();
